Fix for correct handling of vhost-interface.
Contrail compute nodes need to be prepared with Vhost interfaces for
vrouter to run. The vhost0 is created and the IP of eth0 is attached to
vhost0 with the default traffic redirected through the vhost interface.
The create command for vhost cannot work without loading the kernel
module "vrouter". Also, for the contrail_compute recipe to run on the
compute nodes, it must be included from within common_agent and only for
compute nodes with kvm support. The inclusion of the recipe from the
server.rb recipe is redundant since this only runs on the control node.
This patch makes the necessary fixes to handle the above conditions.